extract splitter handler
